requirements.txt
torch
torchvision
opencv-python
matplotlib
Pillow
pycocotools
segmentation-models-pytorch
numpy
seaborn
scikit-learn
albumentations
